fix(groupscreen_scroll): Screen smoothly scrolls header away and tabs stay at the top
